Allow inclusive/exclusive neighborhood selection
This commit changes neighborhood.[ch] to such that only one profile
function is necessary for both the inclusive and exclusive version of
the neighborhood profile.

Change commands.c to work with the new function
interfaces, however it will only select the inclusive version of the
functions. The exclusive option will be available in a future commit.

Update documentation.